Continue working on client.app.home
- Start working on add device screen
- Remove unnecessary not found screen and update failed.dart
- Fix bug during connection process
- Devices are now required to be connected to wifi to use the app

+++ client.app.home/lib/components/button.dart
@@ -5,7 +5,7 @@ class Button extends StatelessWidget {
   const Button({Key key, this.title, this.onPressed, this.width}) : super(key: key);
 
   final String title;
-  final int width;
+  final double width;
   final void Function() onPressed;
 
   @override
